pyproject.toml
[tool.poetry]
authors = ["University of Passau"]
description = "OpenWebSearch Project - RAG"
license = "MIT"
name = "owlergpt"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
beautifulsoup4 = "^4.12.2"
brotlipy = "^0.7.0"
chromadb = "^0.4.14"
click = "^8.1.7"
flask = "^3.0.0"
litellm = "^1.10.4"
pandas = "^2.1.3"
plotly = "^5.18.0"
pypdf = "^4.0.1"
python = "^3.11"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
python-slugify = "^8.0.1"
requests = "^2.31.0"
scikit-learn = "^1.3.2"
sentence-transformers = "^2.2.2"
torch = "2.0.0"
warcio = "^1.7.4"
langchain = "^0.1.7"
opensearch-py = "^2.4.2"
langchain-experimental = "^0.0.54"
langchain-openai = "^0.0.8"
pyarrow = "^15.0.1"
matplotlib = "^3.8.3"
tqdm = "^4.66.2"
seaborn = "^0.13.2"
openai = "^1.19.0"
cohere = "^5.3.0"

[tool.poetry.group.dev.dependencies]
black = "^23.10.1"
flake8 = "^6.1.0"


[[tool.poetry.source]]
name = "pytorch-gpu-src"
url = "https://download.pytorch.org/whl/cu113"
priority = "explicit"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.black]
line-length = 100
